MusicLib: Add comments && clean code && move data classes in separated package
