+++ app/src/main/java/com/app/music_app/note_player/instruments/VirtualPiano.kt
@@ -1,10 +1,10 @@
 package com.app.music_app.note_player.instruments
 
 import com.app.music_app.note_player.interfaces.AbstractInstrument
-import com.musiclib.Alteration
+import com.musiclib.notes.data.Alteration
 import com.musiclib.notes.Note
-import com.musiclib.notes.NoteName
-import com.musiclib.notes.NoteRange
+import com.musiclib.notes.data.NoteName
+import com.musiclib.notes.data.NoteRange
 
 
 // Piano notes: https://github.com/fuhton/piano-mp3
