+++ app/src/test/java/com/app/music_app/VirtualPianoTest.kt
@@ -1,9 +1,9 @@
 package com.app.music_app
 
 import com.app.music_app.note_player.instruments.VirtualPiano
-import com.musiclib.Alteration
+import com.musiclib.notes.data.Alteration
 import com.musiclib.notes.Note
-import com.musiclib.notes.NoteName
+import com.musiclib.notes.data.NoteName
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import kotlin.test.assertFailsWith
