+++ musiclib/src/main/java/com/musiclib/notes/MelodyNote.kt
@@ -1,6 +1,9 @@
 package com.musiclib.notes
 
-import com.musiclib.Alteration
+import com.musiclib.notes.data.Alteration
+import com.musiclib.notes.data.NoteDuration
+import com.musiclib.notes.data.NoteName
+import com.musiclib.notes.data.NoteVolume
 import com.musiclib.notes.interfaces.MelodyNote
 
 /**
