--- musiclib/src/main/java/com/musiclib/notes/NoteVolume.kt
@@ -1,8 +1,0 @@
-package com.musiclib.notes
-
-enum class NoteVolume(val num: Int) {
-    Fortissimo(4), // ff - очень громко
-    Forte(3),      // f - громко
-    Piano(2),      // p - тихо
-    Pianissimo(1)  // pp - очень тихо
-}
