+++ musiclib/src/main/java/com/musiclib/notes/Pause.kt
@@ -1,5 +1,6 @@
 package com.musiclib.notes
 
+import com.musiclib.notes.data.NoteDuration
 import com.musiclib.notes.interfaces.MusicPause
 
 class Pause(override val duration: NoteDuration) : MusicPause {
