+++ musiclib/src/main/java/com/musiclib/notes/data/NoteVolume.kt
@@ -1,0 +1,11 @@
+package com.musiclib.notes.data
+
+/**
+ * Перечисление возможных оттенков громкости звучания ноты
+ */
+enum class NoteVolume() {
+    Fortissimo, // ff - очень громко
+    Forte,      // f - громко
+    Piano,      // p - тихо
+    Pianissimo  // pp - очень тихо
+}
