+++ musiclib/src/main/java/com/musiclib/notes/interfaces/MelodyNote.kt
@@ -1,7 +1,8 @@
 package com.musiclib.notes.interfaces
 
-import com.musiclib.notes.NoteVolume
+import com.musiclib.notes.data.NoteVolume
 
+/** Представляет собой ноту с дополнительными характеристиками звука */
 interface MelodyNote : BasicNote, MusicPause {
     /** Динамика */
     val volume: NoteVolume
