+++ musiclib/src/main/java/com/musiclib/notes/interfaces/MusicPause.kt
@@ -1,6 +1,6 @@
 package com.musiclib.notes.interfaces
 
-import com.musiclib.notes.NoteDuration
+import com.musiclib.notes.data.NoteDuration
 
 /** Представляет собой музыкальную паузу */
 interface MusicPause {
