+++ musiclib/src/test/kotlin/musiclib/test/IntervalTest.kt
@@ -1,9 +1,9 @@
 package musiclib.test
 
-import com.musiclib.Alteration
+import com.musiclib.notes.data.Alteration
 import com.musiclib.intervals.Interval
 import com.musiclib.notes.Note
-import com.musiclib.notes.NoteName
+import com.musiclib.notes.data.NoteName
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
